requirements.txt
fastapi
uvicorn
mlflow
streamlit
torch
torchvision
pillow